remove utils out of celery_app.py
